feat(User): update User to use Fid for internal storage

+++ crates/farmap/src/subset.rs
@@ -141,7 +141,7 @@ mod tests {
         fn test_user_count_before_and_after_filter() {
             let users = dummy_data();
             let mut set = create_set(&users);
-            let fid_filter = |user: &User| is_fid(user, 1);
+            let fid_filter = |user: &User| is_fid(user, 1 as u64);
             check_user_count(&set, 2);
             test_filter::check_filter(&mut set, fid_filter);
             check_user_count(&set, 1);
@@ -151,7 +151,7 @@ mod tests {
         fn test_user_count_before_and_after_filter_two() {
             let users = dummy_data();
             let mut set = create_set(&users);
-            let fid_filter = |user: &User| !is_fid(user, 3);
+            let fid_filter = |user: &User| !is_fid(user, 3 as u64);
             check_user_count(&set, 2);
             test_filter::check_filter(&mut set, fid_filter);
             check_user_count(&set, 2);
